requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-genai==1.68.0
python-dotenv==1.2.2
gunicorn==25.3.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
Werkzeug==3.1.7
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6
greenlet==3.3.2
typing_extensions==4.15.0
anyio==4.13.0
httpx==0.28.1
pydantic==2.12.5
requests==2.33.0
sniffio==1.3.1
tenacity==9.1.4
websockets==16.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
